+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## Added
+- service: a more elaborate error message if get request returns an error on unmarshalling json array
+
 ## [8.5.0]
 
 ## Added